added logic in hooks to create inventory logs

+++ models/medicine.model.js
@@ -1,5 +1,6 @@
 import mongoose, { Schema } from "mongoose";
 import { ExpiryRecord } from "./expiryrecord.model.js";
+import { InventoryLog } from "./inventorylog.model.js";
 
 const medicineSchema = Schema(
   {
@@ -36,9 +37,29 @@ const medicineSchema = Schema(
   { timestamps: true }
 );
 
+medicineSchema.pre("findOneAndDelete", function (next) {
+  this.context = this.getOptions().context;
+  next();
+});
+
 medicineSchema.post("findOneAndDelete", async function (doc) {
-  if (doc) {
+  if (doc && this.context?.user) {
+    const userId = this.context.user.userId;
     try {
+      const stockIds = doc.stockInfo;
+      for (let i = 0; i < stockIds.length; i++) {
+        const deletedExpiryRecord = await ExpiryRecord.findByIdAndDelete(
+          stockIds[i]
+        );
+
+        await InventoryLog.create({
+          medicineId: deletedExpiryRecord.medicineId,
+          updatedBy: userId,
+          batchNumber: deletedExpiryRecord.batchNumber,
+          expiryDate: deletedExpiryRecord.expiryDate,
+          quantity: -1 * deletedExpiryRecord.quantity,
+        });
+      }
       await ExpiryRecord.deleteMany({
         _id: { $in: doc.stockInfo },
       });
